util/process_test.go
package util

import (
	"testing"
)

// TestRunCommand tests the RunCommand function for successfully running an external command.
// Note: This test might fail if the command being tested is not available or if there are issues with the environment.
func TestRunCommand(t *testing.T) {
	// Define a command to test
	cmdName := "echo"
	cmdArgs := []string{"Hello, World!"}

	// Run the command using RunCommand function
	err := RunCommand(cmdName, cmdArgs...)
	if err != nil {
		t.Errorf("RunCommand returned an error: %v", err)
	}
}

// TestRunCommandFailure tests the RunCommand function for handling command start failure.
func TestRunCommandFailure(t *testing.T) {
	// Define a command that will fail to test error handling
	cmdName := "/path/to/nonexistent/command"

	// Run the command using RunCommand function
	err := RunCommand(cmdName)
	if err == nil {
		t.Errorf("RunCommand did not return an error for a non-existent command")
	}
}

// TestRunCommandLifecycle tests the RunCommand function for ensuring child processes have the same lifecycle as the parent.
// This test might be complex to implement and might require mocking or special test conditions.
// It is recommended to manually test this behavior by inspecting how the process group is handled during the test.
func TestRunCommandLifecycle(t *testing.T) {
	// NOTE(lingxuan): TODO.
	// This test requires manual inspection or special test conditions to verify the process group lifecycle.
	t.Skip("Test requires manual inspection or special test conditions to verify the process group lifecycle")
}
